requirements.txt
# Web framework
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
jinja2==3.1.5
aiofiles==24.1.0

# ML / Deep Learning
tensorflow==2.15.1
keras==2.15.0
opencv-python==4.10.0.84

# Data processing
numpy==1.26.4
pandas==2.2.3
scipy==1.13.1
scikit-image==0.23.2
imageio==2.34.2

# Visualization
matplotlib==3.9.2

# Utilities
Pillow==11.3.0
